+++ shop/src/main/java/max/shop/controller/adminController/ItemController.java
@@ -3,8 +3,8 @@ package max.shop.controller.adminController;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import max.shop.domain.Item;
-import max.shop.dto.request.ItemCreateForm;
-import max.shop.dto.request.UpdateItemDto;
+import max.shop.dto.request.item.ItemCreateForm;
+import max.shop.dto.request.item.UpdateItemDto;
 import max.shop.service.item.ItemServiceImpl;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
