+++ shop/src/main/java/max/shop/domain/User.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 import max.shop.domain.embed.Address;
 import max.shop.domain.time.BaseTimeEntity;
 import max.shop.domain.type.Gender;
-import max.shop.dto.request.UserRegisterForm;
+import max.shop.dto.request.user.UserRegisterForm;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
