+++ shop/src/main/java/max/shop/service/item/ItemService.java
@@ -1,8 +1,8 @@
 package max.shop.service.item;
 
 import max.shop.domain.Item;
-import max.shop.dto.request.ItemCreateForm;
-import max.shop.dto.request.UpdateItemDto;
+import max.shop.dto.request.item.ItemCreateForm;
+import max.shop.dto.request.item.UpdateItemDto;
 
 import java.util.List;
