+++ shop/src/main/java/max/shop/service/item/ItemServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import max.shop.common.exception.ItemNotFoundException;
 import max.shop.domain.Item;
-import max.shop.dto.request.ItemCreateForm;
-import max.shop.dto.request.UpdateItemDto;
+import max.shop.dto.request.item.ItemCreateForm;
+import max.shop.dto.request.item.UpdateItemDto;
 import max.shop.repository.ItemRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
