+++ shop/src/test/java/max/shop/service/OrderServiceTest.java
@@ -8,8 +8,8 @@ import max.shop.domain.User;
 import max.shop.domain.embed.Address;
 import max.shop.domain.type.Gender;
 import max.shop.domain.type.OrderStatus;
-import max.shop.dto.request.ItemCreateForm;
-import max.shop.dto.request.UserRegisterForm;
+import max.shop.dto.request.item.ItemCreateForm;
+import max.shop.dto.request.user.UserRegisterForm;
 import max.shop.repository.OrderRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
